Extract ValidatorCumulativeVoting component
